Clean up first part of JsonApiXSpecBuilders
