+++ processor/src/main/java/com/infinum/jsonapix/processor/specs/JsonApiXSpecBuilder.kt
@@ -3,11 +3,6 @@ package com.infinum.jsonapix.processor.specs
 import com.infinum.jsonapix.core.JsonApiX
 import com.infinum.jsonapix.core.common.JsonApiConstants
 import com.infinum.jsonapix.core.common.JsonApiConstants.Prefix.withName
-import com.infinum.jsonapix.core.resources.Error
-import com.infinum.jsonapix.core.resources.Links
-import com.infinum.jsonapix.core.resources.Meta
-import com.infinum.jsonapix.core.resources.ResourceObject
-import com.squareup.kotlinpoet.ANY
 import com.squareup.kotlinpoet.AnnotationSpec
 import com.squareup.kotlinpoet.ClassName
 import com.squareup.kotlinpoet.FileSpec
@@ -16,19 +11,16 @@ import com.squareup.kotlinpoet.KModifier
 import com.squareup.kotlinpoet.ParameterSpec
 import com.squareup.kotlinpoet.ParameterizedTypeName.Companion.parameterizedBy
 import com.squareup.kotlinpoet.PropertySpec
-import com.squareup.kotlinpoet.TypeName
 import com.squareup.kotlinpoet.TypeSpec
 import com.squareup.kotlinpoet.asClassName
-import kotlinx.serialization.Contextual
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.Transient
 
-internal object JsonApiXSpecBuilder {
+internal object JsonApiXSpecBuilder : BaseJsonApiXSpecBuilder() {
 
     private val serializableClassName = Serializable::class.asClassName()
 
-    @SuppressWarnings("LongMethod")
-    fun build(
+    override fun build(
         className: ClassName,
         type: String,
         metaClassName: ClassName?
@@ -39,8 +31,8 @@ internal object JsonApiXSpecBuilder {
             JsonApiConstants.Prefix.RESOURCE_OBJECT.withName(className.simpleName)
         )
 
-        val properties = mutableListOf<PropertySpec>()
-        val params = mutableListOf<ParameterSpec>()
+        val properties = getBasePropertySpecs(metaClassName).toMutableList()
+        val params = getBaseParamSpecs(metaClassName).toMutableList()
 
         params.add(
             ParameterSpec.builder(
@@ -50,50 +42,6 @@ internal object JsonApiXSpecBuilder {
         )
 
         properties.add(dataProperty(resourceObjectClassName))
-
-        params.add(
-            Specs.getNullParamSpec(
-                JsonApiConstants.Keys.INCLUDED,
-                List::class.asClassName().parameterizedBy(
-                    ResourceObject::class.asClassName()
-                        .parameterizedBy(getAnnotatedAnyType())
-                ).copy(nullable = true)
-            )
-        )
-        properties.add(
-            Specs.getNullPropertySpec(
-                JsonApiConstants.Keys.INCLUDED,
-                List::class.asClassName().parameterizedBy(
-                    ResourceObject::class.asClassName()
-                        .parameterizedBy(getAnnotatedAnyType())
-                ).copy(nullable = true)
-            )
-        )
-
-        params.add(
-            ParameterSpec.builder(
-                JsonApiConstants.Keys.ERRORS,
-                List::class.parameterizedBy(Error::class)
-                    .copy(nullable = true)
-            ).defaultValue("%L", "null")
-                .build()
-        )
-
-        properties.add(errorsProperty())
-
-        properties.add(
-            Specs.getNamedPropertySpec(Links::class.asClassName(), JsonApiConstants.Keys.LINKS, true)
-        )
-        params.add(Specs.getNamedParamSpec(Links::class.asClassName(), JsonApiConstants.Keys.LINKS, true))
-
-        params.add(
-            ParameterSpec.builder(
-                JsonApiConstants.Keys.META,
-                metaClassName?.copy(nullable = true) ?: Meta::class.asClassName().copy(nullable = true)
-            ).defaultValue("%L", "null").build()
-        )
-
-        properties.add(metaProperty(metaClassName))
 
         return FileSpec.builder(className.packageName, generatedName)
             .addImport(
@@ -123,26 +71,13 @@ internal object JsonApiXSpecBuilder {
             .build()
     }
 
-    private fun getAnnotatedAnyType(): TypeName {
-        val contextual = AnnotationSpec.builder(Contextual::class).build()
-        return ANY.copy(annotations = ANY.annotations + contextual)
-    }
-
     private fun dataProperty(resourceObject: ClassName): PropertySpec = PropertySpec.builder(
-        JsonApiConstants.Keys.DATA, resourceObject
+        JsonApiConstants.Keys.DATA,
+        resourceObject
     ).addAnnotation(
         Specs.getSerialNameSpec(JsonApiConstants.Keys.DATA)
     )
         .initializer(JsonApiConstants.Keys.DATA).addModifiers(KModifier.OVERRIDE)
-        .build()
-
-    private fun errorsProperty(): PropertySpec = PropertySpec.builder(
-        JsonApiConstants.Keys.ERRORS,
-        List::class.parameterizedBy(Error::class).copy(nullable = true),
-        KModifier.OVERRIDE
-    )
-        .addAnnotation(Specs.getSerialNameSpec(JsonApiConstants.Keys.ERRORS))
-        .initializer(JsonApiConstants.Keys.ERRORS)
         .build()
 
     private fun originalProperty(
@@ -155,17 +90,5 @@ internal object JsonApiXSpecBuilder {
         ).addAnnotation(AnnotationSpec.builder(Transient::class.asClassName()).build())
 
         return builder.initializer(codeString).build()
-    }
-
-    private fun metaProperty(
-        metaClassName: ClassName?
-    ): PropertySpec {
-        return PropertySpec.builder(
-            JsonApiConstants.Keys.META,
-            metaClassName?.copy(nullable = true) ?: Meta::class.asClassName().copy(nullable = true),
-            KModifier.OVERRIDE
-        ).addAnnotation(Specs.getSerialNameSpec(JsonApiConstants.Keys.META))
-            .initializer(JsonApiConstants.Keys.META)
-            .build()
     }
 }
